backend/pyproject.toml
[project]
name = "ai-py"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.135.1",
    "langchain>=1.2.10",
    "langchain-openai>=1.1.10",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.41.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
